+++ lib/src/utils/board_option.dart
@@ -1,3 +1,6 @@
+import 'dart:math';
+
+import 'package:board_datetime_picker/src/utils/datetime_util.dart';
 import 'package:flutter/material.dart';
 
 import '../parts/item.dart';
@@ -7,8 +10,10 @@ class BoardPickerItemOption {
   BoardPickerItemOption({
     required this.type,
     required this.focusNode,
-    required this.list,
-    required this.selected,
+    required this.map,
+    required this.selectedIndex,
+    required this.minimumDate,
+    required this.maximumDate,
   });
 
   /// [DateType] year, month, day, hour, minute
@@ -17,70 +22,87 @@ class BoardPickerItemOption {
   /// TextField [FocusNode].
   final FocusNode focusNode;
 
-  /// Picker item list.
-  List<int> list;
+  /// Picker item map.
+  Map<int, int> map;
 
   /// Selected item for list.
-  int selected;
+  int selectedIndex;
 
   /// Keys for date item widget
   final stateKey = GlobalKey<ItemWidgetState>();
 
   /// Minimum year that can be specified
-  static const int minYear = 1900;
+  final DateTime minimumDate;
+
+  /// Maximum year that can be specified
+
+  final DateTime maximumDate;
 
   /// Constractor
-  factory BoardPickerItemOption.init(DateType type, DateTime date) {
-    List<int> list = [];
+  factory BoardPickerItemOption.init(
+    DateType type,
+    DateTime date,
+    DateTime? minimum,
+    DateTime? maximum,
+  ) {
+    Map<int, int> map = {};
     int selected;
+
+    // Define specified minimum and maximum dates
+    final mi = minimum ?? DateTimeUtil.defaultMinDate;
+    final ma = maximum ?? DateTimeUtil.defaultMaxDate;
+
     switch (type) {
       case DateType.year:
-        return BoardPickerItemOption.year(date);
+        return BoardPickerItemOption.year(date, minimum, maximum);
       case DateType.month:
-        for (var i = 1; i <= 12; i++) {
-          list.add(i);
-        }
-        selected = date.month;
+        map = minmaxList(DateType.month, date, mi, ma);
+        selected = indexFromValue(date.month, map);
         break;
       case DateType.day:
-        for (var i = 1; i <= 31; i++) {
-          list.add(i);
-        }
-        selected = date.day;
+        map = minmaxList(DateType.day, date, mi, ma);
+        selected = indexFromValue(date.day, map);
         break;
       case DateType.hour:
-        for (var i = 0; i <= 23; i++) {
-          list.add(i);
-        }
-        selected = date.hour;
+        map = minmaxList(DateType.hour, date, mi, ma);
+        selected = indexFromValue(date.hour, map);
         break;
       case DateType.minute:
-        for (var i = 0; i <= 59; i++) {
-          list.add(i);
-        }
-        selected = date.minute;
+        map = minmaxList(DateType.minute, date, mi, ma);
+        selected = indexFromValue(date.minute, map);
         break;
     }
 
     return BoardPickerItemOption(
       focusNode: FocusNode(),
-      list: list,
+      map: map,
       type: type,
-      selected: selected,
+      selectedIndex: selected,
+      minimumDate: mi,
+      maximumDate: ma,
     );
   }
 
   /// Constractor for year item
-  factory BoardPickerItemOption.year(DateTime date) {
-    List<int> list = [];
-    for (var i = minYear; i < 2099; i++) {
-      list.add(i);
-    }
+  factory BoardPickerItemOption.year(
+    DateTime date,
+    DateTime? minimum,
+    DateTime? maximum,
+  ) {
+    final minY = minimum?.year ?? DateTimeUtil.minimumYear;
+
+    // Define specified minimum and maximum dates
+    final mi = minimum ?? DateTime(DateTimeUtil.minimumYear, 1, 1, 0, 0);
+    final ma = maximum ?? DateTime(DateTimeUtil.maximumYear, 12, 31, 23, 59);
+
+    final map = minmaxList(DateType.year, date, mi, ma);
     return BoardPickerItemOption(
       focusNode: FocusNode(),
-      list: list,
+      map: map,
       type: DateType.year,
-      selected: date.year,
+      selectedIndex: indexFromValue(max(date.year, minY), map),
+      minimumDate: mi,
+      maximumDate: ma,
     );
   }
 
@@ -100,64 +122,44 @@ class BoardPickerItemOption {
     return 1;
   }
 
-  int getIndex({int? index}) {
-    switch (type) {
-      case DateType.year:
-        return (index ?? selected) - minYear;
-      case DateType.month:
-      case DateType.day:
-        return (index ?? selected) - 1;
-      default:
-        return index ?? selected;
-    }
-  }
-
-  int getValueFromIndex(int index) {
-    switch (type) {
-      case DateType.year:
-        return minYear + index;
-      case DateType.month:
-      case DateType.day:
-        return index + 1;
-      default:
-        return index;
-    }
-  }
-
-  void updateList(int maxDay) {
-    List<int> tmp = [];
-    for (var i = 1; i <= maxDay; i++) {
-      tmp.add(i);
-    }
-    list = tmp;
-
-    if (selected >= list.length) {
-      selected = list.last;
-    }
-
-    stateKey.currentState?.updateState(list);
-  }
-
   void changeDate(DateTime date) {
     switch (type) {
       case DateType.year:
-        selected = date.year;
+        selectedIndex = getIndexFromValue(date.year) ?? 0;
         break;
       case DateType.month:
-        selected = date.month;
+        selectedIndex = getIndexFromValue(date.month) ?? 0;
         break;
       case DateType.day:
-        selected = date.day;
+        selectedIndex = getIndexFromValue(date.day) ?? 0;
         break;
       case DateType.hour:
-        selected = date.hour;
+        selectedIndex = getIndexFromValue(date.hour) ?? 0;
         break;
       case DateType.minute:
-        selected = date.minute;
+        selectedIndex = getIndexFromValue(date.minute) ?? 0;
         break;
     }
-    stateKey.currentState?.toAnimateChange(selected, button: true);
+    stateKey.currentState?.toAnimateChange(selectedIndex, button: true);
   }
+
+  /// Get the map index from the value and return it
+  int? getIndexFromValue(int val) {
+    for (final index in map.keys) {
+      if (map[index] == val) return index;
+    }
+    return null;
+  }
+
+  static int indexFromValue(int val, Map<int, int> map) {
+    for (final index in map.keys) {
+      if (map[index] == val) return index;
+    }
+    return 0;
+  }
+
+  /// Get the currently selected value
+  int get value => map[selectedIndex]!;
 
   /// input content check
   void checkInputField() {
@@ -165,7 +167,7 @@ class BoardPickerItemOption {
     if (text != null) {
       try {
         final data = int.parse(text);
-        if (list.contains(data)) return;
+        if (map.values.contains(data)) return;
       } catch (_) {}
     }
 
@@ -176,11 +178,118 @@ class BoardPickerItemOption {
     stateKey.currentState?.textController.text = date.year.toString();
   }
 
+  void updateList(DateTime date) {
+    //  Retrieve existing values
+    final tmp = value;
+    // Generate new maps
+    map = minmaxList(type, date, minimumDate, maximumDate);
+    _updateState(tmp, date);
+  }
+
+  void updateDayMap(int maxDay, DateTime newDate) {
+    int minDay = 1;
+    int maxDay = 31;
+    if (newDate.isMinimum(minimumDate, DateType.month)) {
+      minDay = minimumDate.day;
+    }
+    if (newDate.isMaximum(maximumDate, DateType.month)) {
+      maxDay = min(maxDay, maximumDate.day);
+    }
+
+    //  Retrieve existing values
+    final tmp = value;
+
+    Map<int, int> newMap = {};
+    int index = 0;
+    for (var i = minDay; i <= maxDay; i++) {
+      newMap[index] = i;
+      index++;
+    }
+    map = newMap;
+    _updateState(tmp, newDate);
+  }
+
+  void _updateState(int tmpValue, DateTime date) {
+    // Get the index of the value that was selected
+    // before the update and update it to that value
+    final index = getIndexFromValue(tmpValue);
+    if (index != null) {
+      selectedIndex = index;
+    } else if (map.values.first > tmpValue) {
+      selectedIndex = 0;
+    } else if (selectedIndex >= map.keys.length) {
+      selectedIndex = map.keys.last;
+    }
+    stateKey.currentState?.updateState(map, selectedIndex);
+  }
+
+  static Map<int, int> minmaxList(
+    DateType dt,
+    DateTime date,
+    DateTime minimum,
+    DateTime maximum,
+  ) {
+    Map<int, int> createMap(int start, int end) {
+      Map<int, int> map = {};
+      int index = 0;
+      for (var i = start; i <= end; i++) {
+        map[index] = i;
+        index++;
+      }
+      return map;
+    }
+
+    switch (dt) {
+      case DateType.year:
+        return createMap(minimum.year, maximum.year);
+      case DateType.month:
+        int minMonth = 1;
+        int maxMonth = 12;
+        if (date.isMinimum(minimum, DateType.year)) {
+          minMonth = minimum.month;
+        }
+        if (date.isMaximum(maximum, DateType.year)) {
+          maxMonth = maximum.month;
+        }
+        return createMap(minMonth, maxMonth);
+      case DateType.day:
+        int minDay = 1;
+        int maxDay = 31;
+        if (date.isMinimum(minimum, DateType.month)) {
+          minDay = minimum.day;
+        }
+        if (date.isMaximum(maximum, DateType.month)) {
+          maxDay = maximum.day;
+        }
+        return createMap(minDay, maxDay);
+      case DateType.hour:
+        int minHour = 0;
+        int maxHour = 23;
+        if (date.isMinimum(minimum, DateType.day)) {
+          minHour = minimum.hour;
+        }
+        if (date.isMaximum(maximum, DateType.day)) {
+          maxHour = maximum.hour;
+        }
+        return createMap(minHour, maxHour);
+      case DateType.minute:
+        int minMinute = 0;
+        int maxMinute = 59;
+        if (date.isMinimum(minimum, DateType.hour)) {
+          minMinute = minimum.minute;
+        }
+        if (date.isMaximum(maximum, DateType.hour)) {
+          maxMinute = maximum.minute;
+        }
+        return createMap(minMinute, maxMinute);
+    }
+  }
+
   DateTime calcDate(DateTime date, {int? index}) {
     switch (type) {
       case DateType.year:
         return DateTime(
-          (index ?? selected),
+          map[index ?? selectedIndex]!,
           date.month,
           date.day,
           date.hour,
@@ -189,7 +298,7 @@ class BoardPickerItemOption {
       case DateType.month:
         return DateTime(
           date.year,
-          (index ?? selected),
+          map[index ?? selectedIndex]!,
           date.day,
           date.hour,
           date.minute,
@@ -198,7 +307,7 @@ class BoardPickerItemOption {
         return DateTime(
           date.year,
           date.month,
-          (index ?? selected),
+          map[index ?? selectedIndex]!,
           date.hour,
           date.minute,
         );
@@ -207,7 +316,7 @@ class BoardPickerItemOption {
           date.year,
           date.month,
           date.day,
-          (index ?? selected),
+          map[index ?? selectedIndex]!,
           date.minute,
         );
       case DateType.minute:
@@ -216,7 +325,7 @@ class BoardPickerItemOption {
           date.month,
           date.day,
           date.hour,
-          (index ?? selected),
+          map[index ?? selectedIndex]!,
         );
     }
   }