+++ lib/src/utils/datetime_util.dart
@@ -3,6 +3,18 @@ import 'package:board_datetime_picker/src/utils/board_option.dart';
 import 'board_enum.dart';
 
 class DateTimeUtil {
+  ///. Default Minimum Date Year
+  static const int minimumYear = 1970;
+
+  /// Default Maximum Date Year
+  static const int maximumYear = 2050;
+
+  ///. Default Minimum Date
+  static DateTime defaultMinDate = DateTime(minimumYear, 1, 1, 0, 0);
+
+  /// Default Maximum Date
+  static DateTime defaultMaxDate = DateTime(maximumYear, 12, 31, 23, 59);
+
   /// Update the date when the year or month is changed.
   /// Therefore, get the last date that exists in year and month.
   static int? getExistsDate(
@@ -14,8 +26,11 @@ class DateTimeUtil {
       return null;
     }
 
-    var year = options.firstWhere((x) => x.type == DateType.year).selected;
-    var month = options.firstWhere((x) => x.type == DateType.month).selected;
+    final yearOption = options.firstWhere((x) => x.type == DateType.year);
+    final monthOption = options.firstWhere((x) => x.type == DateType.month);
+
+    var year = yearOption.value;
+    var month = monthOption.value;
 
     if (selected.type == DateType.year) {
       year = val;
@@ -29,9 +44,125 @@ class DateTimeUtil {
 
     return date.day;
   }
+
+  /// If minimum and maximum dates are specified,
+  /// check whether they are within the range and return values that fall within the range.
+  static DateTime rangeDate(
+    DateTime date,
+    DateTime? minimumDate,
+    DateTime? maximumDate,
+  ) {
+    DateTime newVal = date;
+    if (minimumDate != null && date.isBefore(minimumDate)) {
+      newVal = minimumDate;
+    } else if (maximumDate != null && date.isAfter(maximumDate)) {
+      newVal = maximumDate;
+    }
+    return newVal;
+  }
 }
 
 extension DateTimeExtension on DateTime {
+  bool isMinimum(DateTime date, DateType dt, {bool equal = true}) {
+    bool operator(a, b) {
+      if (equal) {
+        return a <= b;
+      }
+      return a < b;
+    }
+
+    switch (dt) {
+      case DateType.year:
+        return operator(year, date.year);
+      case DateType.month:
+        if (year < date.year) return true;
+        return year <= date.year && operator(month, date.month);
+      case DateType.day:
+        if (year < date.year) return true;
+        if (year <= date.year && month < date.month) return true;
+        return year <= date.year &&
+            month <= date.month &&
+            operator(day, date.day);
+      case DateType.hour:
+        if (year < date.year) return true;
+        if (year <= date.year && month < date.month) return true;
+        if (year <= date.year && month <= date.month && day < date.day) {
+          return true;
+        }
+        return year <= date.year &&
+            month <= date.month &&
+            day <= date.day &&
+            operator(hour, date.hour);
+      case DateType.minute:
+        if (year < date.year) return true;
+        if (year <= date.year && month < date.month) return true;
+        if (year <= date.year && month <= date.month && day < date.day) {
+          return true;
+        }
+        if (year <= date.year &&
+            month <= date.month &&
+            day <= date.day &&
+            hour < date.hour) {
+          return true;
+        }
+        return year <= date.year &&
+            month <= date.month &&
+            day <= date.day &&
+            hour <= date.hour &&
+            operator(minute, date.minute);
+    }
+  }
+
+  bool isMaximum(DateTime date, DateType dt, {bool equal = true}) {
+    bool operator(a, b) {
+      if (equal) {
+        return a >= b;
+      }
+      return a > b;
+    }
+
+    switch (dt) {
+      case DateType.year:
+        return operator(year, date.year);
+      case DateType.month:
+        if (year > date.year) return true;
+        return year >= date.year && operator(month, date.month);
+      case DateType.day:
+        if (year > date.year) return true;
+        if (year >= date.year && month > date.month) return true;
+        return year >= date.year &&
+            month >= date.month &&
+            operator(day, date.day);
+      case DateType.hour:
+        if (year > date.year) return true;
+        if (year >= date.year && month > date.month) return true;
+        if (year >= date.year && month >= date.month && day > date.day) {
+          return true;
+        }
+        return year >= date.year &&
+            month >= date.month &&
+            day >= date.day &&
+            operator(hour, date.hour);
+      case DateType.minute:
+        if (year > date.year) return true;
+        if (year >= date.year && month > date.month) return true;
+        if (year >= date.year && month >= date.month && day > date.day) {
+          return true;
+        }
+        if (year >= date.year &&
+            month >= date.month &&
+            day >= date.day &&
+            hour > date.hour) {
+          return true;
+        }
+        return year >= date.year &&
+            month >= date.month &&
+            day >= date.day &&
+            hour >= date.hour &&
+            operator(minute, date.minute);
+    }
+  }
+
   bool compareDate(DateTime d1) {
     return d1.year == year && d1.month == month && d1.day == day;
   }
@@ -44,8 +175,8 @@ extension DateTimeExtension on DateTime {
 
       if (month >= 12) {
         final x = nextMonth % 12;
-        nextYear += x;
-        nextMonth = nextMonth ~/ 12;
+        nextYear += nextMonth ~/ 12;
+        nextMonth = x;
       }
       date = DateTime(nextYear, nextMonth, 1);
     } else if (diff < 0) {
@@ -57,5 +188,10 @@ extension DateTimeExtension on DateTime {
       date = x0;
     }
     return date;
+  }
+
+  /// Check if the date is within the specified range
+  bool isWithinRange(DateTime minimum, DateTime maximum) {
+    return isAfter(minimum) && isBefore(maximum);
   }
 }