+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.3
+* Allows minimum and maximum dates to be specified
+* Added `openPicker` that can be displayed without arguments when open.
+* [Bug fix] Fixed date misalignment issue with calendar paging
+
 ## 1.0.2
 * Changed to allow specification of language data to be displayed
 * [Bug fix] Fixed problem with date buttons not becoming selected